+++ tests/reduct/lifecycle_api_test.cc
@@ -15,7 +15,7 @@ IClient::LifecycleSettings DefaultSettings() {
       .type = IClient::LifecycleType::kDelete,
       .bucket = "test_bucket_1",
       .entries = {"entry-1"},
-      .max_age = "1h",
+      .older_than = "1h",
       .interval = "10m",
       .mode = IClient::LifecycleMode::kEnabled,
   };
@@ -35,6 +35,7 @@ TEST_CASE("reduct::Client should get list of lifecycles", "[lifecycle_api][1_20]
 TEST_CASE("reduct::Client should create a lifecycle", "[lifecycle_api][1_20]") {
   Fixture ctx;
   auto settings = DefaultSettings();
+  settings.type = IClient::LifecycleType::kCompress;
 
   auto err = ctx.client->CreateLifecycle("test_lifecycle", settings);
   REQUIRE(err == Error::kOk);
@@ -63,7 +64,7 @@ TEST_CASE("reduct::Client should update a lifecycle", "[lifecycle_api][1_20]") {
   REQUIRE(err == Error::kOk);
 
   settings.entries = {"entry-2"};
-  settings.max_age = "2h";
+  settings.older_than = "2h";
   settings.interval = "20m";
   err = ctx.client->UpdateLifecycle("test_lifecycle", settings);
   REQUIRE(err == Error::kOk);