+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add lifecycle policy API support, [PR-128](https://github.com/reductstore/reduct-cpp/pull/128)
+
 ## 1.19.1 - 2026-04-21
 
 ### Fixed